requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-Migrate==4.0.7
python-dotenv==1.0.1
pywebview==4.4.1